requirements.txt
rich
questionary
python-dotenv
jinja2